fix: ファイルを分割し、usecase層から呼ばれることを想定したコードに変更
